cli, db: print uint8_t arguments as decimal in error messages
std::cerr << unsigned_char writes the byte as a glyph, not as a
number. The "must be in the range 0 to 255." and "ascii no N"
diagnostics relied on the latter, so they emitted byte 0xFF as an
invalid-UTF-8 glyph instead of "255", and small ASCII codes as
unprintable controls instead of decimals. Cast both numeric uses to
unsigned int at the call site — matching the explicit-cast pattern
already used at cli.cc:328 — so the messages render as intended.

Repro before:
    $ swarm -d -1 /dev/null
    Error: Illegal ... must be in the range 0 to <0xff>.
After:
    Error: Illegal ... must be in the range 0 to 255.

+++ src/cli.cc
@@ -533,7 +533,7 @@ auto validate_clustering(struct Parameters const & parameters) -> void {
   static constexpr auto uint8_max = std::numeric_limits<uint8_t>::max();
   if ((parameters.opt_differences < 0) or (parameters.opt_differences > uint8_max)) {
     fatal("Illegal number of differences specified with -d or --differences, "
-          "must be in the range 0 to ", uint8_max, ".");
+          "must be in the range 0 to ", static_cast<unsigned int>(uint8_max), ".");
   }
 }
